flytter strategier til egen pakke og regel resultat til egne filer
