+++ formidling/src/main/java/no/nav/ung/sak/formidling/vedtak/regler/strategy/EndringBarnetilleggStrategy.java
@@ -1,36 +1,34 @@
-package no.nav.ung.sak.formidling.vedtak.regler;
+package no.nav.ung.sak.formidling.vedtak.regler.strategy;
 
 import jakarta.enterprise.context.Dependent;
 import jakarta.inject.Inject;
 import no.nav.fpsak.tidsserie.LocalDateTimeline;
 import no.nav.ung.kodeverk.dokument.DokumentMalType;
 import no.nav.ung.sak.behandlingslager.behandling.Behandling;
-import no.nav.ung.sak.formidling.innhold.EndringProgramPeriodeInnholdBygger;
+import no.nav.ung.sak.formidling.innhold.EndringBarnetilleggInnholdBygger;
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultat;
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultatType;
 
 @Dependent
-public final class EndringStartdatoStrategy implements VedtaksbrevInnholdbyggerStrategy {
+public final class EndringBarnetilleggStrategy implements VedtaksbrevInnholdbyggerStrategy {
 
-    private final EndringProgramPeriodeInnholdBygger endringProgramPeriodeInnholdBygger;
+    private final EndringBarnetilleggInnholdBygger endringBarnetilleggInnholdBygger;
 
     @Inject
-    public EndringStartdatoStrategy(EndringProgramPeriodeInnholdBygger endringProgramPeriodeInnholdBygger) {
-        this.endringProgramPeriodeInnholdBygger = endringProgramPeriodeInnholdBygger;
+    public EndringBarnetilleggStrategy(EndringBarnetilleggInnholdBygger endringBarnetilleggInnholdBygger) {
+        this.endringBarnetilleggInnholdBygger = endringBarnetilleggInnholdBygger;
     }
 
     @Override
     public VedtaksbrevStrategyResultat evaluer(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
-        return new VedtaksbrevStrategyResultat(DokumentMalType.ENDRING_PROGRAMPERIODE, endringProgramPeriodeInnholdBygger, "Automatisk brev ved endring av startdato", null);
+        return new VedtaksbrevStrategyResultat(DokumentMalType.ENDRING_BARNETILLEGG, endringBarnetilleggInnholdBygger, "Automatisk brev ved fødsel av barn.", null);
     }
 
     @Override
     public boolean skalEvaluere(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
         var resultatInfo = VedtaksbrevInnholdbyggerStrategy.tilResultatInfo(detaljertResultat);
         var resultater = new ResultatHelper(resultatInfo);
-        return resultater
-            .utenom(DetaljertResultatType.INNVILGET_UTEN_ÅRSAK)
-            .innholderBare(DetaljertResultatType.ENDRING_STARTDATO);
+        return resultater.innholder(DetaljertResultatType.ENDRING_BARN_FØDSEL);
     }
 
 }
