+++ formidling/src/main/java/no/nav/ung/sak/formidling/vedtak/regler/strategy/EndringHøySatsStrategy.java
@@ -1,34 +1,34 @@
-package no.nav.ung.sak.formidling.vedtak.regler;
+package no.nav.ung.sak.formidling.vedtak.regler.strategy;
 
 import jakarta.enterprise.context.Dependent;
 import jakarta.inject.Inject;
 import no.nav.fpsak.tidsserie.LocalDateTimeline;
 import no.nav.ung.kodeverk.dokument.DokumentMalType;
 import no.nav.ung.sak.behandlingslager.behandling.Behandling;
-import no.nav.ung.sak.formidling.innhold.EndringBarnetilleggInnholdBygger;
+import no.nav.ung.sak.formidling.innhold.EndringHøySatsInnholdBygger;
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultat;
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultatType;
 
 @Dependent
-public final class EndringBarnetilleggStrategy implements VedtaksbrevInnholdbyggerStrategy {
+public final class EndringHøySatsStrategy implements VedtaksbrevInnholdbyggerStrategy {
 
-    private final EndringBarnetilleggInnholdBygger endringBarnetilleggInnholdBygger;
+    private final EndringHøySatsInnholdBygger endringHøySatsInnholdBygger;
 
     @Inject
-    public EndringBarnetilleggStrategy(EndringBarnetilleggInnholdBygger endringBarnetilleggInnholdBygger) {
-        this.endringBarnetilleggInnholdBygger = endringBarnetilleggInnholdBygger;
+    public EndringHøySatsStrategy(EndringHøySatsInnholdBygger endringHøySatsInnholdBygger) {
+        this.endringHøySatsInnholdBygger = endringHøySatsInnholdBygger;
     }
 
     @Override
     public VedtaksbrevStrategyResultat evaluer(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
-        return new VedtaksbrevStrategyResultat(DokumentMalType.ENDRING_BARNETILLEGG, endringBarnetilleggInnholdBygger, "Automatisk brev ved fødsel av barn.", null);
+        return new VedtaksbrevStrategyResultat(DokumentMalType.ENDRING_HØY_SATS, endringHøySatsInnholdBygger, "Automatisk brev ved endring til høy sats.", null);
     }
 
     @Override
     public boolean skalEvaluere(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
         var resultatInfo = VedtaksbrevInnholdbyggerStrategy.tilResultatInfo(detaljertResultat);
         var resultater = new ResultatHelper(resultatInfo);
-        return resultater.innholder(DetaljertResultatType.ENDRING_BARN_FØDSEL);
+        return resultater.innholderBare(DetaljertResultatType.ENDRING_ØKT_SATS);
     }
 
 }
