+++ formidling/src/main/java/no/nav/ung/sak/formidling/vedtak/regler/strategy/EndringStartdatoStrategy.java
@@ -1,34 +1,36 @@
-package no.nav.ung.sak.formidling.vedtak.regler;
+package no.nav.ung.sak.formidling.vedtak.regler.strategy;
 
 import jakarta.enterprise.context.Dependent;
 import jakarta.inject.Inject;
 import no.nav.fpsak.tidsserie.LocalDateTimeline;
 import no.nav.ung.kodeverk.dokument.DokumentMalType;
 import no.nav.ung.sak.behandlingslager.behandling.Behandling;
-import no.nav.ung.sak.formidling.innhold.EndringHøySatsInnholdBygger;
+import no.nav.ung.sak.formidling.innhold.EndringProgramPeriodeInnholdBygger;
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultat;
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultatType;
 
 @Dependent
-public final class EndringHøySatsStrategy implements VedtaksbrevInnholdbyggerStrategy {
+public final class EndringStartdatoStrategy implements VedtaksbrevInnholdbyggerStrategy {
 
-    private final EndringHøySatsInnholdBygger endringHøySatsInnholdBygger;
+    private final EndringProgramPeriodeInnholdBygger endringProgramPeriodeInnholdBygger;
 
     @Inject
-    public EndringHøySatsStrategy(EndringHøySatsInnholdBygger endringHøySatsInnholdBygger) {
-        this.endringHøySatsInnholdBygger = endringHøySatsInnholdBygger;
+    public EndringStartdatoStrategy(EndringProgramPeriodeInnholdBygger endringProgramPeriodeInnholdBygger) {
+        this.endringProgramPeriodeInnholdBygger = endringProgramPeriodeInnholdBygger;
     }
 
     @Override
     public VedtaksbrevStrategyResultat evaluer(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
-        return new VedtaksbrevStrategyResultat(DokumentMalType.ENDRING_HØY_SATS, endringHøySatsInnholdBygger, "Automatisk brev ved endring til høy sats.", null);
+        return new VedtaksbrevStrategyResultat(DokumentMalType.ENDRING_PROGRAMPERIODE, endringProgramPeriodeInnholdBygger, "Automatisk brev ved endring av startdato", null);
     }
 
     @Override
     public boolean skalEvaluere(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
         var resultatInfo = VedtaksbrevInnholdbyggerStrategy.tilResultatInfo(detaljertResultat);
         var resultater = new ResultatHelper(resultatInfo);
-        return resultater.innholderBare(DetaljertResultatType.ENDRING_ØKT_SATS);
+        return resultater
+            .utenom(DetaljertResultatType.INNVILGET_UTEN_ÅRSAK)
+            .innholderBare(DetaljertResultatType.ENDRING_STARTDATO);
     }
 
 }
