+++ formidling/src/main/java/no/nav/ung/sak/formidling/vedtak/regler/strategy/VedtaksbrevInnholdbyggerStrategy.java
@@ -1,4 +1,4 @@
-package no.nav.ung.sak.formidling.vedtak.regler;
+package no.nav.ung.sak.formidling.vedtak.regler.strategy;
 
 import no.nav.fpsak.tidsserie.LocalDateTimeline;
 import no.nav.ung.sak.behandlingslager.behandling.Behandling;
