+++ formidling/src/main/java/no/nav/ung/sak/formidling/vedtak/regler/strategy/VedtaksbrevStrategyResultat.java
@@ -1,7 +1,8 @@
-package no.nav.ung.sak.formidling.vedtak.regler;
+package no.nav.ung.sak.formidling.vedtak.regler.strategy;
 
 import no.nav.ung.kodeverk.dokument.DokumentMalType;
 import no.nav.ung.sak.formidling.innhold.VedtaksbrevInnholdBygger;
+import no.nav.ung.sak.formidling.vedtak.regler.IngenBrevÅrsakType;
 
 public record VedtaksbrevStrategyResultat(
     DokumentMalType dokumentMalType,
